Moving event callback out of entry point manager
